refactor: Notice, Settings 컴포넌트 ChatPlugin 으로 이동
